feat: reduced size of parameters in callback

+++ drivers/inc/display.h
@@ -52,9 +52,9 @@ void print_backspace();
 void init_display();
 uint16_t get_screen_width();
 uint16_t get_screen_height();
-void put_pixel(int x, int y, uint32_t color);
-void put_line(int x, int y, int lenght, uint32_t color);
-void draw_rectangle(int x, int y, int w, int h, uint32_t color);
+void put_pixel(uint16_t x, uint16_t y, uint32_t color);
+void put_line(uint16_t x, uint16_t y, uint16_t lenght, uint32_t color);
+void draw_rectangle(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint32_t color);
 void fill_screen(uint32_t color);
