+++ drivers/src/display.c
@@ -112,7 +112,7 @@ uint16_t get_screen_height() {
     return mode_info->height;
 }
 
-void put_pixel(int x, int y, uint32_t color) {
+void put_pixel(uint16_t x, uint16_t y, uint32_t color) {
     uint8_t* framebuffer = (uint8_t*)(uintptr_t)mode_info->framebuffer;
 
     uint32_t offset = y * mode_info->pitch + x * 3;
@@ -122,13 +122,13 @@ void put_pixel(int x, int y, uint32_t color) {
     framebuffer[offset + 2] = (color >> 16) & 0xFF;
 }
 
-void put_line(int x, int y, int lenght, uint32_t color) {
+void put_line(uint16_t x, uint16_t y, uint16_t lenght, uint32_t color) {
     for (int i=0; i<lenght; i++) {
         put_pixel(x+i, y, color);
     }
 }
 
-void draw_rectangle(int x, int y, int w, int h, uint32_t color) {
+void draw_rectangle(uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint32_t color) {
     for (int dy = y; dy < y + h; dy++) {
         for (int dx = x; dx < x + w; dx++) {
             put_pixel(dx, dy, color);
